Moved default colors.
